backend/requirements.txt
blinker==1.6.2
click==8.1.6
distlib==0.3.7
dnspython==2.4.1
filelock==3.12.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT==0.3.2
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
mysqlclient==2.1.1
platformdirs==3.10.0
PyJWT==1.4.2
pymongo==4.4.1
PyMySQL==1.1.0
virtualenv==20.24.2
Werkzeug==2.3.6
